Remove manage cyles, rename initiate cycle
